refactor(web): split assistant-side-menu into subcomponents
assistant-side-menu.tsx owns section composition and menu wiring; its
three self-contained subtrees live in their own files: the built-in nav
block (assistant cluster + pinned apps) in side-menu-built-in-nav.tsx,
the overlay's floating bottom column with its height measurement in
side-menu-overlay-bottom-column.tsx, and the collapsed rail's flyout
icon list in collapsed-rail-sections.tsx. 872 -> 634 lines, mechanical
moves only.
